Prevent directives from being overwritten
